[acpisvc] support acpi events
Add the acpi_enable_event API. This call creates a port associated with
the ACPI handle, and ACPI events are delivered on that handle.

There is a limitation to the current implementation. If there are multiple
ACPI handles to the same object, acpi_enable_event will only succeed for
the first handle that calls the API.

Change-Id: I6d7abfcae7aecd61a4ad8c419b159ebb879d836f